:sparkles: (RUMM-770): Tweak React Native Android code generation

+++ tests/gen_reactnative/test_interface/expected/android/src/main/java/com/datadog/reactnative/Setter.kt
@@ -39,8 +39,8 @@ class Setter(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule
      * @param value A long param
      */
     @ReactMethod
-    fun setLong(value: Long, promise: Promise) {
-        nativeInstance.setLong(value)
+    fun setLong(value: Double, promise: Promise) {
+        nativeInstance.setLong(value.toLong())
         promise.resolve(null)
     }
